Move settings tab content to seperate components
